+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- [Development] Discontinues use of FakeRedis in favor of real Redis for testing.
+
 ## [1.1.0] - 2025-05-13
 
 ### Changed